examples/sms-home-stack/requirements.txt
RPi.GPIO==0.7.0
sms-stack